backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
python-telegram-bot==20.3
aiohttp==3.9.1
requests==2.31.0
httpx==0.24.1
